Use the same parameter order as numpy.{lin,log}space}.

+++ AnaTools/interface/CommonUtils.h
@@ -224,8 +224,8 @@ namespace anatools
 
   template<class T> bool isMatchedToTriggerObject (const edm::Event &, const edm::TriggerResults &, const T &, const vector<pat::TriggerObjectStandAlone> &, const string &, const string &, const double dR = 0.1);
 
-  void logSpace (const unsigned, const double, const double, vector<double> &);
-  void linSpace (const unsigned, const double, const double, vector<double> &);
+  void logSpace (const double, const double, const unsigned, vector<double> &);
+  void linSpace (const double, const double, const unsigned, vector<double> &);
 }
 
 /**
